Revise playRound function and create game function that iterates over playRound 5 times
In the playRound function, I changed the return values to integers to make it easier to handle when using the return value in the game function. In the game function, I log the choices of the player and computer, and log the current score. After the 5th iteration, I log the results of the game.